fix: :bug: edited migration book
- alter isbn column for 13 digits;
- alter value column for 2 places;

+++ database/migrations/2024_05_15_024948_create_books_table.php
@@ -13,8 +13,8 @@ return new class () extends Migration {
         Schema::create('books', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100);
-            $table->integer('isbn')->nullable();
-            $table->decimal('value', 10, 8)->nullable();
+            $table->decimal('isbn', 13, 0)->nullable();
+            $table->decimal('value', 10, 2)->nullable();
             $table->timestamps();
         });
     }
